Extract unified diff handling to a separate function
